requirements/dev_requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --output-file requirements/dev_requirements.txt requirements/requirements.txt requirements/dev_requirements.in
asgiref==3.11.0
    # via
    #   -r requirements/requirements.txt
    #   django
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
django==5.2.8
    # via
    #   -r requirements/requirements.txt
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
django-stubs==5.2.8
    # via
    #   -r requirements/dev_requirements.in
    #   djangorestframework-stubs
django-stubs-ext==5.2.8
    # via django-stubs
djangorestframework==3.16.1
    # via -r requirements/requirements.txt
djangorestframework-stubs==3.16.6
    # via -r requirements/dev_requirements.in
idna==3.11
    # via requests
python-decouple==3.8
    # via -r requirements/requirements.txt
python-dotenv==1.2.1
    # via -r requirements/requirements.txt
requests==2.32.5
    # via djangorestframework-stubs
ruff==0.14.7
    # via -r requirements/dev_requirements.in
sqlparse==0.5.4
    # via
    #   -r requirements/requirements.txt
    #   django
types-pyyaml==6.0.12.20250915
    # via
    #   django-stubs
    #   djangorestframework-stubs
types-requests==2.32.4.20250913
    # via djangorestframework-stubs
typing-extensions==4.15.0
    # via
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
urllib3==2.5.0
    # via
    #   requests
    #   types-requests
